move Job and TimeStamp classes
